SINGA-89  Add Docker support
Docker is a Linux container which is fast to deploy and incurs small overhead. A Docker container functions like a light-weight virtual machine and runs in its own isolated environment. There are at least two benefits in building adding Docker support for SINGA:

+ Out of the box deployment of SINGA: the image, once built, contains the complete environment necessary to start SINGA.

+ Light-weight development and testing environment for distributed features. The image can be used to set up a test-bed consisting of many independent nodes (each node has its own IP address), without the need for cluster hardware.

We add to Github a Dockerfile specifying how to build SINGA image, from which the user can construct the image by executing:

sudo docker build -t singa/base .

The user can choose another name beside singa/base . The build process can take a long time, but it needs to be done only at one host and the result image can then be copied to other hosts.

We also add another Dockerfile which adds Mesos and Hadoop ontop of SINGA. This is closely related to SINGA-11. The image created from this Dockerfile is used to set up the distributed test bed.

See the README.md for more details of the images and how to use them.

IMPORTANT We assume that every host has Docker running. Two nodes on the epiC cluster (`ciidaa-c18` and `ciidaa-c19`) are set up with Docker and the pre-built images.

IMPORTANT For now, the Dockerfile for building Mesos+Hadoop+SINGA will pull the latest from ug93tad's (Anh's) SINGA-11 branch. This pull step is removed after SINGA-11 is merged to the project's master branch.