SINGA-25 - Setup glog output path
Make the log_dir an optional configuration in cluster.conf.
If this field is not set, then log is output to /tmp/ using glog's default file name.
If this field is set, then log is output to log_dir folder with the file name as <model name>-<log level>-<date time>