SINGA-28 Fix a bug from toplogy sort of Graph
There is a bug in the Sort function of graph.cc,
which would push one node into the 'tmp' queue for
multiple times if this node has multiple source nodes.

Solution: use an unordered_set to store all nodes in the queue;
push a node only when it is not in the 'tmp' queue.