SINGA-32 Implement synchronous training framework
Fix a bug from InitLocalParam() of Worker class.
One worker owns one Param if the param's data blob is not shared from other workers.
Previously, a Worker would not send Get request for one Param if it owns this Param.
But it may not init the Param locally because it is not the first group in a group
set which subscribe to the same server group.

To fix the bug, all workers would send Get requests for Params in its local layers.
There would not extra cost for getting Params owned by the worker (from the first group),
because the get reqest would not be sent (the param version is already the latest).