SINGA-12 Supprt Checkpoint and Restore
The checkpoint is done in the Worker class and controlled by two model configuration fields: checkpoint_after and checkpoint_frequency.
Only do checkpoint for the Params owning the param values from the first group.
The name, version and values of one Param are dumped onto disk (the path is <workspace>/checkpoint/step<training step>-worker<worker id>.bin).
It is possible that the snapshot is separated into mutiple files because the neural net is partitioned into multi workers.

The checkpoint files can be used:
application 1: to restore (resume) the training by setting the command line argument -resume = true.
The Resume function of the Trainer will find the files for the latest snapshot and add them to the model.conf's checkpoint filed.
It also set the model config's step field to the snapshot step (extracted from file name).

application 2: as the pretraining result of another model. Users have to config the new model's checkpoint field to add the paths of all checkpoint files of the same snapshot.

The Worker's InitLocalParam will init Param's from checkpoint files if available. Otherwise it randomly init them using user configured init method.
It matches the Param objects based on name. If the Param is not configured with a name, NeuralNet class will automatically create one for it based on the name
of the layer to which the Param belongs. In this case,
for application 2, users have to either configure the names of the new model params carefully to
match the names of params from the pre-trained model
for application 1, the worker-server topology cannot be changed.

Restore for params which are partitioned due to model partitioning is not supported. Because if the pre-training is done using 2 workers, while the new model is trained with 3 workers,
then the same original param is partitioned in different ways and hence cannot be matched.