SINGA-9 Add Support for Restricted Boltzman Machine (RBM) model
Implement the Contrastive Divergence (CD) algorithm to train RBM model.
We have implemented a BPWorker to run the Back-Propagation algorithm. To implement the CD algorithm, we follow the same way to create a CDWorker whose RunOneBatch function controls the logic of the CD algorithm, including positive phase, negative phase and computing gradient phase.