SINGA-55 Refactor main.cc and singa.h
A Driver class is implemented for initializing SINGA including parsing job ID,
and registering built-in subclasses of Layer, Updater, Worker and Param.
May use it to init MPI if using MPI as the message passing lib.

The main.cc file is updated to provide an example main func.

Update the GaussianSqrtFanin init method for Param (ref SINGA-58).