SINGA-61 Support user defined classes
To support user defined classes, we need to set a type field for each class in the configuration.
E.g., the LayerProto has a type field. But this type field can only be set to built-in layer types.
This commit adds a user_type field for user defined layers. E.g.,

    LayerProto {
      optional string user_type = 10;
    }

A helper function Layer::Create(const LayerProto&) is added to create a Layer object based on configured type.

Similar updates are applied for other classes, namely, Param, Worker, Updater.