SINGA-71 Fix a buf from repeated averaging gradients from workers
The Trainer::HandleUpdate() averaged the gradients from local workers
from the same group. But the average is also conducted in the
Updater::Update() function. In other words, the gradients were averaged
twice which made them smaller than normal values, and the learning
slower.

The bug is fixed by removing the averaging operation in
Trainer::HandleUpdate()