SINGA-72 Minor updates to be consisten with documentation
Update install.sh to make OpenBLAS with ONLY_CBLAS=1, which removes the
dependency on gfortran. Thanks to ZhangHao for reporting this solution.

Corret the configuration for rnnlm example (replacing test with
validation).

Add the license of cpplint.py in LICENSE file.

Add the solution of openblas installation error into README.md.