SINGA-85 Add functions for extracting features and test new data
* to test pre-trained model,
  1. add checkpoint_path in job configuration to the checkpoint files
  2. setup input layer properly to read test data
  3. run singa with -test argument (e.g., ./bin/singa-run.sh -conf JOB_CONF  -test

* to extract features of new data,
  1. add checkpoint_path in job configuration to the checkpoint files
  2. setup input layer properly to read raw data
  3. add outputl layers (e.g., RecordOutputLayer or CSVOutputLayer) after the layer whose data you want to extract.
  4. run singa with -test argument (e.g., ./bin/singa-run.sh -conf JOB_CONF  -test

A test net (i.e., kTest phase is passed to creat the net)  will be create to test performance and extract features.
job.conf of cifar10 example contains instructions for extracting features and testing performance with checkpoint files.

Rename ProtoRecordLayer and CSVRecordLayer to RecordInputLayer and CSVInputLayer respectively.
Add RecordOutputLayer and CSVOutpuLayer for outputing data as RecordProto and CSV format.