SINGA-113 Model/Hybrid Partition Support
Fix a bug caused by the mutable_grad(Layer*) function of SplitLayer and SlicLayer.
They return the grad blob based on dest layer's partition_id.
The bug we encountered:
the slice layer is connected to a BridgeSrcLayer and a ConcateLayer, both with
partition id = 0 (the partition id of the BridgeDstLayer connected from BridgeSrcLayer is 1);
during backward propagation, both BridgeSrcLayer and ConcateLayer get
the set same grad blob of the SliceLayer, hence this blob is overwritten and the other grad Blob is empty.