SINGA-169 Add base Layer class for V1.0
Adapt caffe.proto to create layer.proto.
As long as we support the similar set of layers as other systems, e.g., caffe/torch,
we would have the similar hyper-parameters. If we define our own layer.proto, we only create new names for the same hyper-parameters.
Considering caffe has implemented a wide range of layers, it would be conveient to adapt based caffe.proto.
More over, we can then run caffe's model configurations or pre-trained model in SINGA with little changes.

We change the 'parameter' to 'conf' to distinguish parameter and hyper-parameter (use configuration).
Some fields are not used, but their tags are reserved.
We will later write a converter program/script to run caffe models.

The base Layer class API is drafted.