SINGA-171 - Create CppDevice and CudaDevice
Rename Device subclasses based on the programming language and hardware,
e.g., CppCPU indicates the device is a CPU which runs cpp code, CudaGPU
indicates the device is a NvidiaGPU which runs cuda code, and CudaCPU
indicates the device is a CPU which uses cuda to malloc and free pinned
memory for the CudaGPU.

Corrspondingly, we rename the lib namepace to lang. and Device type()
to lang().