SINGA-194 Add a Platform singleton
Add the Platform class whose methods are all static.
It includes methods for query hardware GPUs, e.g., num of gpus, memory of each gpu.
It also creats a vector of singa::Device, e.g., CudaGPU.
NOTE:
If multiple CudaGPU devices are created, and they all use CnMemPool,
then they must share the same CnMemPool instance (otherwise there would
be problems in desctructing the CnMemPool).
It is preferred to use Platform to create CudaGPU devices which would
avoid the above problem.

Updated the APIs for Device and DeviceMemPool.
The GetAllocateMem is not accurate (which is larger than actually requested memory size).

NOTE:
The swig would report errors if USE_CUDA is OFF as the CudaGPU is not avaiable.
TODO(wangwei) resolve this problem in later PR for python setup.