SINGA-226 Add parallel training on a single machine for singa v1.0
This commit implements a updater class for parallel training
in local-cpu and local-gpu mode. (The mode specification described
in https://issues.apache.org/jira/browse/SINGA-226)

Updater class is a wrapper of Optimizer class. It controls the
communication pattern among workers. When initializing Updater,
the user needs to provide the total number of workers and where
the Updater does aggregation and computation.

File changed descibed as follows:
* Put a new folder named cifar-parallel under example, containing
  the single machine multi-gpu parallel training example.
* Replace Optimizer pointer in feed_forward_net class with Updater
  pointer, since Updater is a wrapper of Optimizer. So the compile()
  method is changed accordingly.
* Add a helper function TrainThread() method in feed_forward_net to
  launch a training thread.
* Adapt alexnet.cc original cifar10 example to support the new compile()
* Fixed a bug in memory.cc which happens during initialzing cnmem
  memory pool