SINGA-259 Add maven pom file for building java classes
This ticket would refactor some API in *.i files and create the pom.xml
to use maven to compile the java classes. Java unittest is also enabled
with a simple example.

To test this ticket,
1. compile singa with java  (cuda and cudnn are optional) in `build/`

        cmake -DUSE_JAVA=ON -DUSE_PYTHON=OFF ..
        make

2. goto build/java and run the unit tests

    export LD_LIBRARY_PATH=<path to
    build>/java/lib:$LD_LIBRARY_PATH
    mvn test

The folder structure for the java project is

    java/
    pom.xml
    lib/libsinga_wrap.so
    src/
      main/java/org/apache/singa/
                              swig/
                              proto/
      test/org/apache/singa/

Fixed a bug in tensor.py::from_numpy() which handled int array incorrectly.