SINGA-264 Extend the FeedForwardNet to accept multiple inputs
Extend FeedForwardNet to support multiple input tensors and output tensors.
The input variable x, of train(x, y), forward(x), predict(x), evaluate(x)
could be a single tensor or a dictionary: layer name -> a single tensor or tensor list.
The key is the name of the layer to feed the input data.

The output of out=forward(x, output), would be a single tensor or a dictionary:
layer name -> a single tensor or a tensor list. The key is the name of
the layer to get the output values, e.g, the net has multiple layers
whose outgoing degree is 0. By configuring the argument
'output' as a list of layer names, we can get values of those layers in
'out'.

Passed unittests.