SINGA-409 [Singa 1.2.0] Basic `singa-cpu` import throws error
It is due to the version mismatch of Protobuf.
In SINGA-396, we updated the protobuf version to be >=3.2.0. And the
singa package is built using 3.2.0.
However, 3.6.1 is installed in your OS. In other words, the runtime env
and the building env mismatch.
In this commit, we fix the version for both building and runtime to
3.6.1.