SINGA-478 Add the __itruediv__ for python 3
We need to add  _itruediv_ in tensor.py because the original _idiv_ is not supported by python 3 anymore.
 
To understand the problem, let's study the following code first:

from singa import tensor
from singa import device
import numpy as np

Y = np.ones(shape=[10],dtype=np.float32) * 10.0
y = tensor.from_numpy(Y)
y.to_device(device.get_default_device())

def divide(y):
    y /= 10

divide(y)
print(tensor.to_numpy(y))


Without adding the _itruediv_ function, the result is as follows, which means that the /= operation is not in place: 
[10. 10. 10. 10. 10. 10. 10. 10. 10. 10.]
After adding the _itruediv_ function, the result is as follows, which means that the /= operation is in place:
[1. 1. 1. 1. 1. 1. 1. 1. 1. 1.]
This is because the _idiv_ operation is for python 2, while _itruediv_ is for python 3. Therefore, if we do not add the _itruediv_ operator in tensor.py, it just uses a default operation which is not in place.