Test anaconda upload using github ci
The anaconda upload token is a secret on github (nudles).
If the secret is not available, no uploading.
Note: github ci cannot access the secret directly within the `if` check;
you have to assign it to a env var first and then do if check;
better do the upload on push (not on pull request).