Update Githut CI for conda build
Add the CI to run on MacOS, including building, testing pysinga and
uploading the package to anaconda if the token is available.

Fix the permission issue https://github.com/conda/conda/issues/7267
and the sdk issue.
The sdk is fixed to the latest version (in conda_build_config.yaml).
The python is fixed to 3.6 to be compatible with colab.